Crate linfa_linalg

Provides pure-Rust implementations of linear algebra routines for ndarray without depending on external LAPACK/BLAS libraries.

Eliminating BLAS dependencies

If this crate is being used as a BLAS-less replacement for ndarray-linalg, make sure to remove ndarray-linalg from the entire dependency tree of your crate. This is because ndarray-linalg, even as a transitive dependency, forces ndarray to be built with the blas feature, which forces all matrix multiplications to rely on a BLAS backend. This leads to linker errors if no BLAS backend is specified.

Modules

bidiagonal

Compact bidiagonal decomposition for matrices

cholesky

Cholesky decomposition on symmetric positive definite matrices.

eigh

Eigendecomposition for symmetric square matrices

lobpcg

Locally Optimal Block Preconditioned Conjugate Gradient (LOBPCG) is a matrix-free method for finding the large (or smallest) eigenvalues and the corresponding eigenvectors of a symmetric eigenvalue problem

norm

Norm of vectors

qr

QR decomposition of rectangular matrices.

reflection

svd

Compact singular-value decomposition of matrices

triangular

Traits for creating and manipulating triangular matrices

tridiagonal

Tridiagonal decomposition of a symmetric matrix

Enums

LinalgError

Order

Find largest or smallest eigenvalues

Type Definitions

Result