Trait linfa_linalg::eigh::EigSort

pub trait EigSort: Sized {
    // Required method
    fn sort_eig(self, order: Order) -> Self;

    // Provided methods
    fn sort_eig_asc(self) -> Self { ... }
    fn sort_eig_desc(self) -> Self { ... }
}

Sorting of eigendecomposition by the eigenvalues.

Panic

Will panic if shape or layout of inputs differ from eigen output, or if input contains NaN.

Required Methods

fn sort_eig(self, order: Order) -> Self

Provided Methods

fn sort_eig_asc(self) -> Self

Sort eigendecomposition by the eigenvalues in ascending order

fn sort_eig_desc(self) -> Self

Sort eigendecomposition by the eigenvalues in descending order