Module linfa_linalg::qr

QR decomposition of rectangular matrices.

Note that the QR decomposition implemented here is “thin”, so Q has dimensions (r, c) and R is (c, c), where r and c are the dimensions of the original matrix.

This module also exports functionality for calculating matrix inverse and the least squares problem.

Structs

QRDecomp

Compact representation of a QR decomposition. Can be used to yield the Q and R matrices or to calculate the inverse or solve a system.

Traits

LeastSquaresQr

Use QR decomposition to calculate least squares by reference. The A matrix is still passed by value.

LeastSquaresQrInto

Use QR decomposition to calculate least squares by value

QR

QR decomposition for matrix by reference

QRInto

QR decomposition for matrix by value