Trait linfa_linalg::svd::SvdSort

pub trait SvdSort: Sized {
    // Required method
    fn sort_svd(self, order: Order) -> Self;

    // Provided methods
    fn sort_svd_asc(self) -> Self { ... }
    fn sort_svd_desc(self) -> Self { ... }
}

Sorting of SVD decomposition by the singular values. Rearranges the columns of U and rows of Vt accordingly.

Panic

Will panic if shape of inputs differs from shape of SVD output, or if input contains NaN.

Required Methods

fn sort_svd(self, order: Order) -> Self

Provided Methods

fn sort_svd_asc(self) -> Self

Sort SVD decomposition by the singular values in ascending order

fn sort_svd_desc(self) -> Self

Sort SVD decomposition by the singular values in descending order